Show old sidebar when vector_tile flag is inactive
Restore the old sidebar facilities tab as a
`NonVectorTileFilterSidebarFacilitiesTab` component so that it can be
displayed when the vector_tile flag is off.

Restore some utility functions (moved into the NonVectorTile...
component) which were required to make the component work, along with
some Redux actions it expected to exist.

Adjust the fetchFacilities function to accept an options object so that
callers can set the page request size. We don't need to make a similar
adjustment to the fetchNextPage function since it uses the URL returned
from the API for its next page of results.