Implement clicking on fishnet grid layer zooms to the cell bounds
This commit activates our fishnet grid layer by zooming to the bounds of any
cell that is clicked if that cell covers an area containing one or more
facilities matching the search parameters.

We implemented this by encoding the z, x, and y of each cell as attributes of
the square polygons in the vector tile and then using the `sphericalmercator`
library in the layer click event handler to convert those tile numbers into a
bounding box.