Add color ramp legend to facility map
The color ramp represents a relative number of facilities that changes as the
zoom level increases so we use the generic labels "fewer" and "more" on the
legend. We extracted the legend color ramp to the constants file so we can use
it both from the legend and the map layer itself.