Fetch all facilities by default on main page
When the vector tile feature is active, fetch all facilities by default
on the main map and sidebar page. Also make the "Facilities" tab the
initially displayed tab on loading the map page if the vector tile
feature is active.

Add guards to retain prior map behavior when the vector tile feature is
not active.