Encode querystring params vector tile cache key
Encode qs params into vector tile cache key sent with client tile
requests. We do this client side because the client is the source of the
request and already knows what filters have been applied -- and because
the server sent cache key only encodes the version number and the
timestamp of the most recently updated facility.

The server already only checks for the presence or absence of a cache
key in the URL path, so this implementation leaves that unchanged. For future CF or S3
based caching, the path will still be set up to "just work" with
requests for specific sets of tiles.