Cache vector tiles for a year
Add a `Cache-Control: max-age=31536000` header to the
vector tile responses. This allows the browser to cache
the tile for a year.

Since we have pretty sophisticated cache invalidation
in place thanks to the cache key, we can effectively
cache the tiles forever and let LRU take care of losing
the old ones.

This header will also be used by Cloudfront to cache the
tiles after the first user request is made.