Configure Cloudfront to respect tile cache
This configuration ensures that Cloudfront caches the tiles
as soon as a user-requested tile is generated.

This forwards headers (which may include a referrer) and
cookies (which may help identify users by their session id)
to the original request.

Don't forward any cookies, as we want the cache to work across
users, and each user will have a different CSRF Token and
Session ID. Also, limit the headers to Referer. These will be
used for request authentication.