Add update location dashboard page
The layout of this page is based on other "two panel" pages, like facility
merging.

We added "Location" to the `DashboardFacilityCard` so the existing coordinates
are visible. This adds the location to every dashboard page that uses the
`DashboardFacilityCard` but showing extra information is not a problem.

The optional notes and contributor fields allow a superuser to record optional
descriptive detail regarding the change in location and optionally specify the
contributor that provided the updated coordinates.