Add liststosources management command
This command will be used as part of the backward compatible migration process.
It creates or updates `Source` objects from `FacilityList` objects and also
updates the `source` foreign key on related `FacilityListItems`.

It is intended that this command be removed after deployed application code is
updated to create and reference `Source` objects.