Adjust SRID transformation to remove visual artifacts
Previously we were generating the hex grid in 4326 so we could
more easily compare the location data to it. This caused visual
artifacts in some zoom levels.

By creating the hex grid in its native SRID of 3857, and only
converting specific instances to 4326 for comparison at the last
step, we ensure that there are no visual artifacts.