Remove fields made redundant by Source
This is the final step of a multi-step migration to move fields from
`FacilityList` to `Source`

  - Remove `facility_list` from `FacilityListItem`. We now use `source`.
  - Remove `contributor`, `is_active`, and `is_public` from `FacilityList`. We
    now use `source`.

We also make the previously nullable `source` field non-null. Code has been
updated to populate it for new data, and the `liststosources` management command
ensured that a source was created and correctly populated for every
`FacilityList`.