Extract match methods to new module / refactor for fast single match
In order to support quickly matching individual list items we needed a way of
holding a trained and indexed model in memory and keeping it updated as the
facility data changed. The new `GazetteerCache` class implements this by holding
a trained model in a class variable and reading from the `HistoricalFacility`
table to determine whether items need to be added or removed from the index.

The `GazetteerCache` includes a threading lock to ensure that only one thread
can ever be updating the cached gazetteer.

Although the management command-based matching uses the new `GazetteerCache` the
actual behavior has not changed, since starting a process from scratch will
result in training a new model, as we have always done.

As part of this refactor we have also addressed a problem with single item CSV
uploads causing the match process run out of memory. We address this by changing
the way models are trained. Instead of using the input file, we use 20% of the
submitted list items as our "dirty" data.