Implement single-item matching
This commit adds the matching and persistence logic to the view function.

The flow of the view is based on the batch processing pipeline, but all three
steps are completed at once. Each stage is wrapped in an individual try block so
we can log unexpected failures similar to the batch process.

The "parse" step is combined with the saving of the initial `Source` and
`FacilityListItem` records, since the data fields are submitted "pre-parsed." We
create `Source` and `FacilityListItem` objects even when the user passes the
`create=false` option because we still want to track the submission of the data
in the `Source` table and the performance of the matching in the
`FacilityListItem.processing_results` field.

We were able to reuse the `save_match_details` function used by the batch
processing. The only change we needed to make was to wrap the model save code
with a check of the `Source.create` property.

We needed to adjust the facility delete method because submitting
high-confidence matches with `create=false` was creating `FacilityListItem` rows
with foreign key references to the `Facility`, but not a corresponding
`FacilityMatch` row.