Train Dedupe model in a thread when the app starts
When testing the `GazetteerCache` with ~20,000 facilities, the time it takes to
train the model was exceeding the gunicorn request timeout, which caused
gunicorn to kill the worker and start a new one.

To work around this, we add a new `startup.run` method that is called when the
app starts up and loads `urls.py`. `urls.py` is loaded when running management
commands, so we first check for an environment variable that is only set when
the app is loaded by gunicorn.

In `startup.run` we call `GazetteerCache.get_latest()` to train the model. We
call this on a separate thread so that the web request that started the
app (most likely a call to /health-check/) will return immediately.

While the model is being trained, requests that use the model will block. To
avoid wait times that would prompt gunicorn to kill the process, we have added
our own timeout handling and return 503 to the client with a message explaining
that they should try back in a few minutes.