Added can view full contributor details flag/group
This permission acts as an override to allow members of the "can submit
private facilities" group to still see contributor details.

It replaces checking if the user has contributed any public data, allowing us to
have more control over the visibility of contributors and to avoid a situation
where a contributor submits thousands of private facilities and a single public
one to get access to the full contributor details.