Update API to allow searching for contributor facility overlap
We added a `combine_contributors` query string argument to all filtering
facilities by multiple contributors to be switched from "OR" to "AND,"
effectively showing the facilities shared between the specified contributors.

While adding support for this to the `filter_by_query_params` method we also
replaced the previous filter by contributor code with a statement that computes
the "IN" query on the server side. The large SQL statements generated by passing
thousands of IDs in a literal "IN" statement were so large that they made
analyzing query performance from SQL slow query logs difficult to impossible.