Use a nested query when filtering by contributor type
Previously we were calculating a list of facility IDs in a separate query and
passing that literal list as an "in" parameter. When there are thousands of
matching facilities this creates queries that so large, in terms of raw text
size, that they break query analysis tools.

In development, the latency of tile requests is unaffected by this change.