Use confirmed matches as training data
Attempts to make subsequent matching of replacements lists require less
intervention from the contributor by adding previously confirmed matches to the
training data.

This change does not appear to help. Adding it as a branch so it can be archived
and returned to if needed.

These are the testing steps:

  * Run ./scripts/resetdb
  * Log in as c8@example.com
  * Confirm matches on http://localhost:6543/lists/8
  * Upload src/django/api/management/commands/facility_lists/geocoded/8.csv and
  replace the previous list.
  * Fully process the uploaded list
    * ./scripts/manage batch_process --list-id 16 --action parse
    * ./scripts/manage batch_process --list-id 16 --action geocode
    * ./scripts/manage batch_process --list-id 16 --action match
  * Browse the results http://localhost:6543/lists/16

Adding the manual matches as training data via `markPairs` does not boost the
match scores. The match score for row index 11 is 0.69 both times.