Zoom To Search can be disabled
The zoom to search functionality is turned on and off by a checkbox on
the map, which is selected by default. Users are able to disable zoom to
search by unchecking the checkbox.

The state of the checkbox is held in the UI reducer. This seemed like
a logical location as it primarily controls the way information is
presented, not the information itself.