Update claims to: 'Parent company/supplier group'
We heard during the recent OECD conference that many facility owners or
factory groups do not understand use of the language "Parent Company"
as it relates to company structure.

This field in the OAR "Claim a Facility" process, and on the claimed
profiles, has been updated to: "Parent Company / Supplier Group".