Show original GPS coordinates
After the GPS coordinates for a facility were updated, they were
being filtered out from the facility details view of 'other locations'.
This could be confusing for users relying on the original GPS
coordinates.

There was a filter being applied to the other locations data intended
to prevent the original location from showing up as both an other and
a current location; however, additional filters at the same point apply
specifically to filtering out the current location, so we can include
the original location only when it is not the current location by
simply removing the original-location specific filter.