Show original GPS coordinates in 'other locations'
The automatically generated GPS coordinates are moved to the "other
locations" section on a facility profile when GPS coordinates
are updated.

The original, automatically generated GPS coordinates previously
disappeared from the public view when an update was made to GPS
coordinates. This could be confusing for users who had used the first
set of coordinates as their reference point for a facility.

Two tests have been updated to reflect the changed number of facilities
being returned after an update.