OAR opens to search tab
When loading the website, the search tab is activated on load.

Previously, the search tab was set as the first active, but was
immediately changed to facilities when the facilities and feature flags
loaded. The facilities tab is now not activated in response to feature
flags loading. The facilities tab is still activated in response to
facilities loading in most cases, but on initial page load the search
tab remains active.