Display number of contributors per type
The endpoint to fetch all contributors now queries the number of
contributors per contributor type, the appends the count to the
display name for each contributor type. This allows the UI to
display the number of contributors for each type under the contributor
type dropdown in search.

This functionality was requested by several users, as currently this
information is only available through the admin reports.