Move search results count to tab bar
The search results count is in the tab bar. The total count of
facilities no longer appears in the search tab.

Currently, the search results count still appears in the facilities tab.

The facility count API (for total facilities) isn't being used by the
UI now. We could potentially simplify the code by removing the
facility count actions and reducer.