Prevents map pan on reset
The map was moving twice on reset - once to reset to the initial
position, and once to zoom to the extent of all facilities.

Now if zoom-to-search is active, the map zooms to the extent of all
facilities on reset; and if zoom-to-search is inactive, the map resets
to the initial position.