Add a contributors with active lists report
The exiting /admin/reports/countributors/ report shows all registered
accounts. We need a version that only lists those contributor accounts
that have submitted facilities.

This adds a contributors-with-active-facilities report.
It has the same output schema as the contributors report and
uses the same filter criteria as the /api/contributors endpoint.