Update dev data fixture to exclude a contributor type
We ran into a bug that only appeared on staging because our development data did
not have any contributor types with zero members, but our real-world data does.
In this commit we modified the fixture generator and used it to generate a new
`contributors.json` file to ensure that one of the contributor types is not
assigned to any contributor.