Prevent list refresh on facility confirm
When working through the Confirm / Reject process, the Confirm button
refreshed the entire page, which made it difficult to do a lot of
confirmations. This only occurred when filtering by 'Potential Matches',
as we were re-fetching the list when filtering by potential matches.

When fetching a list, we usually erase the current list and then add
the search results in to an empty field. Now, specifically when fetching
a list after confirming a facility match, we don't erase the list, only
overwriting it when the new list is returned. This allows the view to
maintain its position in the list.