Include contributors if they have any active facilities
We were previously excluding contributors from the filter list if any of their
items were in an error state. This meant that if a contributor submitted a list
with one successfully processed item and one error item, the contributor would
be filtered out.

To address this we are now checking for any item in a `MATCHED` or
`CONFIRMED_MATCH` status. This has an additional benefit of excluding
contributors from the list while their list is still processing.