Return contributor details from list API and include in CSV download
The complex permission and visibility checks required make fetching contributor
details for each facility slow, so we are cutting the maximum results per page
from 500 to 50.

Contributor names and list names can have commas, so we join them with a
vertical bar in the CSV to make machine parsing of multiple contributors
possible.