If there are multiple high quality matches choose exact string match
It is possible for a submitted item to match multiple facilities with greater
than 80% confidence, which we have previously presented to the contributor as
potential matches, requiring an decision. This commit attempts to reduce the
number of potential matches by considering an exact string match (after removing
extraneous characters) to be a better match than other high-confidence matches.