Add PPE fields to the facility detail page gated by a feature flag
Tracking PPE production may not be a long term feature of the OAR so we add put
the display of PPE data behind a feature flag. This also gives us the ability to
release other features to production before the PPE feature set is finalized.

We refactored the feature flag API so that we no longer have to enumerate the
flags by name, reducing the boilerplate required to create a new feature flag.