Add PPE fields to the facility CSV download
Adding the PPE fields to the CSV required adding the fields to the facility
serializer.

We incorporate the existing feature flag into the front end so that PPE fields
are not returned in the CSV if the feature is disabled. Conditionally changing
the fields returned from a serializer is challenging, so we have opted to not
alter the API responses based on the feature flag.