Update PPE fields when splitting and merging facility matches
The PPE fields on a `Facility` are updated when a `FacilityListItem` is matched
to an existing facility. If that match is split, we need to "revert" the
updates. If a moderator merges a new match, we need apply the same PPE field
updates that are done when a match is made automatically.