Revert PPE when sources or matches are made inactive
When a line item is not longer associated with a facility, we revert the PPE
field values on the `Facility`, following the existing workflow of splitting a
match.

We choose to trigger this logic in overridden `save` methods on `Source` and
`FacilityMatch` so that we will apply the same reversion logic any time a block
of code changes `is_active` from `True` to `False`. The `is_active` flag is
sometimes changed by moderators via the Django admin so we need a solution that
will work even when the change in active status is triggered by code that we do
not control.

To reduce duplication we moved the set of `Q` statements that we use to filter
only facilities or items that have one or more PPE fields filled in to a
constant and we added a method to `Facility` handles the conditional reverting
of fields.