Prevent crashing the parse job if ppe_product_types value too long
If there is a validation error on the `ppe_product_types` array field,
`full_clean` appears to set it to an empty string which then causes `save` to
raise an exception. We work around this by setting the field to an empty list if
it is not None and not a list.