Fix exception when matching to a confirmed match facility via API
In a previous commit we added confirmed matches to our canonical list of
facilities using a "synthetic" ID that does not match a record in the Facility
table and added a `reduce_matches` helper function that can be used to process
the dedupe results to replace these synthetic IDs with real IDs.

Creating a facility match through the API has a different code path than the
batch processing and was not properly handling these synthetic IDs, resulting in
unhandled exceptions.