Rework GazetteerCache to shorten transactions and be more resilliant
Prior to this change the training of the dedupe model happened inside of a
database transaction. With a large list of facilities this training can take
significant amount of time. We don't need to update the database after training
is complete, so we have refactored the code to fetch all the required data and
then close the transaction before starting the training. Prior to this change we
were already running our QuerySets through list comprehensions, so we are not
increasing the memory footprint beyond what was already in use.

We also move the updating of the `_version` property to occur inside the loop
where we process each history record individually. Previously, if we exited the
loop early, records that had already been processed would be reprocessed the
next time we updated the model. We also lost the ability to log the last
good record for debugging.

To avoid getting "stuck" attempting to incrementally update the model, failing
with an error, and then hitting the same incremental update error the next time
through the loop we add an exception handler that attempts to post the exception
to Rollbar for debugging and then rebuilds the model from scratch. This will
likely create more timeout errors in exchange for not failing with the same
errors over and over.