Process match records when updating GazetteerCache
The incrementally updated GazetteerCache was not complete because we were
ignoring confirmed matches, which we add to the canonical list of facilities
indexed by a synthetic ID. To address this we add code to the GazetteerCache to
process `HistoricalFacilityMatch` records using a method similar to the one used
for processing `HistoricalFacilty` records.

We added `logger.debug` statements to make it easier to see the processing
actions at runtime.