Add trigram match on facility name fallback when matching via API
This new feature of the facility matching/creation API addresses problems
experiences by an MSI when attempting to match records with partial or
incomplete addresses.

We use the same 0.50 confidence threshold as our dedupe matcher for consistency.
It is a reasonable starting point and could be adjusted in the future. When
testing with realistic data we were sometimes matching dozens of records so we
limit the responses to the first five, which is reasonable given that we are
sorting by similarity score.

We are making this an opt-in feature by adding a new optional query string
argument so all existing behavior is preserved.