Add postfacilitiescsv.py for exercising the facility match API
This command line script reads a CSV with name, address, and
country_name/country columns and POSTs each line to the facility matching API.
It was designed for load testing the matching API.

For convenience the token and root URL can be passed as flagged arguments or set
as TOKEN and ROOTURL environment variables. The values default to those that
will work with the development environment populated with `resetdb`.

The -s,--stop argument allows running tests with small numbers of requests
without needing to create input CSVs of different lengths.