Apply ST_Transform() when hex_grid table is created (#1109)
The JOIN in the hex grid query ends up getting applied against each record in the hex_grid table multiplied by the number of facilities generated by the WHERE filter. As an example, contributor ID 699 causes the join filter to be applied against 800K+ records.

Applying ST_Transform() when the hex_grid table is created, and then using the resulting wgs84_geom reduces the overcall cost of the large JOIN (a query that was taking 1.9s went down to 500ms). Further, indexing the wgs84_geom column takes the query down to ~200ms.