Directly connect health check to GazetteerCache and remove thread
Under load from facility matching API requests our app service tasks are being
marked as unhealthy and being restarted. We have not been able to determine the
reason why our services have been marked as unhealthy so we are attempting to
simplify the code around the `GazetteerCache` and make failure easier to reason
about.

By adding a custom watchman check that calls `GazetteerCache.get_latest` we
ensure that the task will not be marked as healthy until a gazetteer matcher has
been fully trained.

We added an ECS service health check grace period to allow for model training
time we initial tested with 60 seconds and increased to 180 seconds which was
the point at which the instance would be successfully marked as healthy.

We set the gunicorn timeout to 120 seconds because saw `[CRITICAL] WORKER
TIMEOUT` error messages in the ECS task log output when this was set to a lower
value. 90 matches the timeout value we set in docker-compose.yml for development
but it appears that the ECS tasks may require more time.