Load GazetteerCache when app starts
We want our model training to be part of the initialization of a service tasks
so we add an `AppConfig` subclass to move this process earlier in the Django
"boot up" process.

We copied a check of the `SERVER_SOFTWARE` environment variable from the old
`startup.py` which ensures that `get_latest` is only called when the app is
running in a Gunicorn worker, not when the app was loaded to run a management
command.

We are keeping the custom health check call to `get_latest` because it will help
continuously keep the GazetteerCache "fresh" in cases where a list is uploaded
and processed by our batch workflow.