Make Source a read only field in Django admin for FacilityListItem
The Django admin was reading every `Source` in order to populate a dropdown
list. Our database now has over 100,000 source records so loading these into a
dropdown is not useful and resource intensive. There is not currently a need to
ever change the source for a line item, so it is safe to make it read-only.