Catch and handle certain exceptions thrown in Dedupe unindex
When incrementally updating the model held in the `GazetterCache` we make calls
to the `unindex` method when facilities or facility matches have been deleted.
We were seeing KeyError and AttributeError exceptions logged to Rollbar that
were being raised within Dedupe code that we were not able to reproduce in
development in our initial evaluation.

We have opted to work around the issue rather than dig for a root cause by
catching these exceptions, logging them to Rollbar as warnings, and proceeding
with other incremental updates. As a result, the matcher could return matches
that reference now-deleted facilities, so we have a companion update to the
`match_items` method that validates the facility exists before appending it to
the returned match results.

When writing test coverage for this change we realized that the existing dedupe
matching tests would not run if their order was changed and determined that this
was because the internal state of the `GazetterCache` was persisting between
tests. We added a `tearDown` method to reset the `GazetterCache` and removed the
`interspace` helper, the use of which was preventing successful matches.