Remove calls to the gazetteer `unindex` method in GazetteerCache
We were previously calling `cls._gazetter.unindex` to remove records with a
`history_type` of `-` but it was raising exceptions for which we could not
determine the root cause. We have opted to ignore them and filter out no longer
existing records from the match results, which we implemented in a previous
commit.