Increase app_ecs_grace_period_seconds
We were seeing instances not become healthy within 120 seconds, which put us
into a loop. We assume that we did not catch this because we were using an
experimental value of 6 during the testing of recent changes to the service
startup process and then returned it to value of 2.

To choose a new value we modified the grace period through the console and found
150 still resulted in tasks not reaching a healthy status, but 180 allowed
sufficient time.