Update reports to consider sources and source.create
We introduced the `Source` model to unify list-based contributions and
one-at-a-time contributions from the API. We had not updated our reports to
incorporate this change, and we are now starting to get real data from API
contributions.

In addition to focusing more on `Source` we also need to consider the
`Source.create` property. We allow API users to set `create=false` on their POST
requests which prevents the creation of `Facility` and `FacilityMatch` records.
These "test" requests were inflating the numbers in some of the reports.