Add an admin-only ApiBlock REST API
Adds an admin-only API for ApiBlock records, offering list, retrieve,
and update actions.

In the update handler, we return a 400 if the caller attempts to update
the `grace_limit` to a value that is not greater than `limit` and
`actual`.

We set `active` to false when the `grace_limit` is updated.