Allow filtering the Source admin list by type and contributor
Moving ownership of a public list to a new contributor account after they have
registered is a common administrative activity. Now that the application is
receiving thousands of API submissions, it is impractical to browse the Source
list to find a specific list. Adding filters allows for more quickly finding a
specific record.