Fix search and list view on Django admin user page
Attempting to search for a user was raising an exception because by default it
was attempting to search for `first_name` and `last_name` fields that we do not
have on our User model. Setting `search_fields` resolves this.

We don't use `username` on our User model so setting `list_display` suppresses
this always empty field in the list view.