Add a management command to make CSVs from a CSV download
This tool allows us to generate a collection of CSV files suitable for uploading
by processing a CSV downloaded from the running application.

The `--rowcount` option allows us to generate more rows than exist in the input
by copying random rows and shuffling the words within the name and address
field.

Passing a static value to `random.seed` allows us to create a psudo-random set
of output file content but the output content will be stable across multiple
runs of the command given the same input file and argument values.