Fix vagrant provisioning failure by pinning pip python version
This approach was copied from another project.

The development VM has only Python 2.7 installed, and the default pip was
recently upgraded to only work with Python 3. This pinning ensures that the VM
provisions successfully.

Updates to the Ansible version and base box are not directly related, but we
updated them as well to match a more recent, working project.