Fix List Filtering Bug
Facilities were being incorrectly filtered based on the source id
instead of the list id. This has been corrected.

In addition, the lists were being navigated to from the profile based
on the user id, not the contributor id. The lists are now navigated
to based on their associated contributor id.