Adjust settings so fast refresh works in development
After upgrading to CRA 4 and craco hot module reloading was not working. We
referenced another project that was using fast refresh and copied in settings to
match.

We removed the environment variables from the `start` script in `package.json`
because they duplicate variable set in docker-compose.yml.

Within `docker-compose.yml` we reduced `CHOKIDAR_INTERVAL`, set `working_dir`,
and switched to node 12 to match another project using fast refresh.

On other projects we have kept the versions of react and react-dom in sync, so
we updated the version on `react-dom` in `package.json`.