Add scripts/console
One of the "scripts to rule them all" that we had not yet implemented on this
project. One helpful use of this script (which prompted us to add it to the
scripts folder) is logging into the running Django container and running `pip
freeze` to check the version numbers of installed dependencies.

https://github.com/github/scripts-to-rule-them-all