Add Nonstandard Fields Model
Our previous method of looping through a contributor's facility list
items and parsing them for the contributor's nonstandard fields was
timing out on staging. This adds a NonstandardField model, which gets
created when a user creates a facility via API or creates a list,
allowing us to access the user's nonstandard fields more quickly.