Add full-width embedded map support
Adjusts the embedded map preview and sample code to allow users
to set the embedded map to full width, and implements sensible
defaults when moving between full width and set width.

Per a conversation with Scott, some small changes have been made to the
styling of the size section of the embedded map settings.

As part of updating the iframe code, also fixes a type in the iframe
src url.