Update Map Preview on Style Change
The embedded map styles are fetched by the site on page load. This
means that if the styles change after the site loads, the site's style
isn't updated until the page is refreshed. In order to tell the embedded
map preview iframe that a refresh is required, a timestamp is passed to
the iframe and is updated when the config is successfully
updated in the database.