Skip download log and remove columns when embedded
We need to allow downloads from embedded maps but we cannot require
authentication so we route around the POST call to the download log endpoint in
embed mode.

We removed PPE search from the embedded mode and in this commit we follow that
lead and remove PPE columns from the data download.

When in embed mode we hide information about other contributors on the facility
detail page. In this commit we follow that lead and remove the contributor
column from the data download.

We will not be disabling the vector tile feature flag, so I removed code related
to it rather than updating it.

We struggled to correctly pass props to `mapDispatchToProps` in
`DownloadFacilitiesButton` so I moved the action dispatch declaration into the
body of the render function.