Add helper text to Claim a Facility
About 80% of facility claims (rough estimate) do not include a
website/LinkedIn profile or share a link that does not have the
facility address included. We then manually email the person who
submitted the claim to ask for that additional documentation. To help
encourage people to proactively send that documentation, should they
not have a website or LinkedIn profile that lists their facility
address, we've added some helper text to encourage them to send
additional facility documentation to use via email, to help speed
up the facility claim process.