Conditionally change split operation to revert
In production we encountered a situation where a list item was used to create a
facility but then through some sequence of moderation activities was matched to
a different facility. Attempts to use split function in the dashboard failed
because attempting to create a new facility from the list item violated the
uniqueness constraint on `Facility.created_from`.

To work around this problem we now include `facility_created_by_item` in the
data used to populate the dashboard split tool and conditionally offer a
"revert" option in place of split and mirror that change in the API view to
connect the match to the originally created facility rather than attempting to
make a new facility record.