Only apply the fixed minimum embed height at mobile widths
When the contributor uses the 100% width embed option we were enforcing a
minimum allowable height to avoid creating an iframe that is too short to be
functional. Replacing the `max` function with a media query allows us to
conditionally enforce this minimum only at mobile widths.

We append GUID to the DOM ID we use to attach styles so that there is no chance
of the ID conflicting with another element on the host page.