Use contractual limit tracking
Previously, a user's number of API calls was tied to the calendar year,
so they could theoretically make 50 calls on Dec 31 and another 50 on
Jan 1. Once we get into paying customers, this becomes more of a
problem. The API limits are now tied to the contract start date, rather
than the calendar year. For free-tier users, we use the date of their
first successful request as the start of their term.