Align logout button
The button for the navigation dropdown for logged-in users was out of
line with the links, but is now aligned. Logout, unlike the other items
in the navigation, is a button instead of an anchor, because it doesn't
navigate the user but rather performs an action in response to a click.
This requires it to have different styles, which need to be set using
higher specifity CSS classes in order to override the '!important' tags
on other CSS classes for nav dropdowns.