Fix facility history bug
There are cases where a facility claim history object references
a facility which has been deleted. When fetching the facility history,
a ObjectDoesNotExist exception was being thrown. We now catch this
exception and exclude the broken objects.

Note: In most cases we would use `facility is not None` instead of
catching an exception. In this situation, most likely because of a
`history` related quirk, that doesn't resolve the error.