Add facility index
Adds an indexed table of facilities, incorporating fields which are
utilized by the facilities search. Data such as facility contributors,
which require several relationship traversals to find, is denormalized
into array fields. A GinIndex is used to index the array fields.

Adds a management command to index facilities. To index all facilities,
pass no arguments to the management command. To index one or more
facilities, pass the ids of the facilities to be indexed.